Backend/src/requirements.txt
asgiref==3.3.1
certifi==2018.11.29
chardet==3.0.4

config==0.5.0.post0

dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-crispy-forms==1.7.2
django-filter==2.4.0
django-heroku==0.3.1

djangorestframework==3.12.2


gunicorn==20.0.4

mysql-connector==2.2.9
mysql-connector-python==8.0.22


psycopg2==2.8.6

PyMySQL==0.10.1
pyparsing==2.4.0

python-dateutil==2.8.0
python-decouple==3.4
pytz==2019.1
requests==2.21.0

sqlparse==0.3.0
urllib3==1.24.1
virtualenv==16.7.7

whitenoise==5.2.0
